clean architecture refactoring
