+++ interface_adapters/file_controller.py
@@ -10,7 +10,7 @@ Classes:
           across multiple categories using various scoring methods.
 
 Dependencies:
-    - pandas: For handling data operations on DataFrames.
+    - pandas: For handling infrastructure operations on DataFrames.
     - helpers: Contains utility functions for identifying protected categories in a DataFrame.
     - bias_calculator: Contains functions for calculating individual and overall bias scores.
 
@@ -18,7 +18,7 @@ Usage:
     Create a File instance by providing the file path of a CSV dataset.
     Use instance methods to retrieve bias scores for specific categories or an overall score.
 """
-from analysis import helpers, bias_calculator
+from useCases import bias_calculator, helpers
 import pandas as pd
 
 
@@ -116,26 +116,3 @@ class File:
             return bias_calculator.calculate_score_by_fpr_mean(self.df, category)
         return None
 
-    def get_overall_score(self, method="variance"):
-        """
-        Calculate an overall bias score across multiple categories in the dataset.
-
-        The score is calculated by averaging scores for each category based on the specified method.
-
-        Parameters:
-            method (str): The method used to calculate individual category scores. Options are:
-                          - "variance": Uses variance of FPRs for each category.
-                          - "fpr_mean": Uses mean of FPRs for each category.
-                          - "accuracy": Placeholder for calculating based on accuracy (returns 0 by default).
-
-        Returns:
-            float: An average score between 0 and 10, where higher values indicate lower bias across categories.
-
-        Notes:
-            - The function averages scores calculated for each category, so `calculate_score_by_variance`
-              and `calculate_score_by_fpr_mean` should be defined separately in `bias_calculator`.
-            - By default, the "variance" method is used. If "accuracy" is selected, it will currently return 0
-              until `calculate_score_by_accuracy` is implemented.
-            - Assumes each scoring method returns a score between 0 and 10.
-        """
-        return bias_calculator.calculate_overall_score(self.df, self.categories, method=method)
