+++ tests/file_controller_test.py
@@ -1,10 +1,11 @@
 import unittest
-import file_handler
+from interface_adapters import file_controller
+
 
 class TestFileHandler(unittest.TestCase):
     def test_score(self):
-        file_address = "data/small_data_wonull_has_all_classes.csv"
-        file = file_handler.File(file_address)
+        file_address = "../infrastructure/small_data_wonull_has_all_classes.csv"
+        file = file_controller.File(file_address)
         print(file.score)
         self.assertTrue(1)
