+++ tests/use_cases_tests.py
@@ -1,13 +1,14 @@
-import helpers
 import unittest
 import pandas as pd
+
+from useCases import helpers
 
 
 class TestHelperFunctions (unittest.TestCase):
     def setUp(self):
         """
         Create a sample DataFrame and other setup necessary for the tests.
-        This method is called before each test method runs.
+        This method is called before each tests method runs.
         """
         self.data = {
             'Race': ['White', 'White', 'Black', 'White', 'Asian', 'White', 'Black', 'Asian'],
