+++ view/src/components/Navbar.tsx
