+++ view/src/components/SettingModal.tsx
