+++ view/src/pages/WelcomePage.tsx
