+++ view/src/vite-env.d.ts
